# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TwinTrack API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /twins:
    get:
      operationId: listTwins
      tags: [twins]
      summary: List all twin profiles for the current user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of twin profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Twin"
    post:
      operationId: createTwin
      tags: [twins]
      summary: Create a new twin profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTwinBody"
      responses:
        "201":
          description: Created twin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Twin"
  /twins/{id}:
    get:
      operationId: getTwin
      tags: [twins]
      summary: Get a twin profile by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Twin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Twin"
        "404":
          description: Not found
    patch:
      operationId: updateTwin
      tags: [twins]
      summary: Update a twin profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTwinBody"
      responses:
        "200":
          description: Updated twin profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Twin"
        "404":
          description: Not found
    delete:
      operationId: deleteTwin
      tags: [twins]
      summary: Delete a twin profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- SLEEP TRACKING ---
  /sleep:
    get:
      operationId: listSleepEntries
      tags: [sleep]
      summary: List sleep entries for a twin
      parameters:
        - in: query
          name: twinId
          required: true
          schema:
            type: integer
        - in: query
          name: date
          required: false
          schema:
            type: ["string", "null"]
            description: ISO date string (YYYY-MM-DD), omit for all entries
      responses:
        "200":
          description: List of sleep entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SleepEntry"
    post:
      operationId: createSleepEntry
      tags: [sleep]
      summary: Create a new sleep entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSleepEntryBody"
      responses:
        "201":
          description: Created sleep entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepEntry"
  /sleep/{id}:
    patch:
      operationId: updateSleepEntry
      tags: [sleep]
      summary: Update a sleep entry (e.g. stop timer, add end time)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSleepEntryBody"
      responses:
        "200":
          description: Updated sleep entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepEntry"
        "404":
          description: Not found
    delete:
      operationId: deleteSleepEntry
      tags: [sleep]
      summary: Delete a sleep entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /sleep/summary:
    get:
      operationId: getSleepSummary
      tags: [sleep]
      summary: Get daily and weekly sleep summary for a twin
      parameters:
        - in: query
          name: twinId
          required: true
          schema:
            type: integer
        - in: query
          name: date
          required: true
          schema:
            type: string
            description: ISO date string (YYYY-MM-DD)
      responses:
        "200":
          description: Sleep summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SleepSummary"

  # --- FEEDING TRACKING ---
...
  /videos:
    get:
      operationId: listVideos
      tags: [videos]
      summary: List all videos in the library
      parameters:
        - in: query
          name: category
          required: false
          schema:
            type: ["string", "null"]
        - in: query
          name: search
          required: false
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
    post:
      operationId: createVideo
      tags: [videos]
      summary: Add a video to the library (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVideoBody"
      responses:
        "201":
          description: Created video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
  /videos/{id}:
    get:
      operationId: getVideo
      tags: [videos]
      summary: Get a video by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        "404":
          description: Not found
  /videos/{id}/bookmark:
    post:
      operationId: bookmarkVideo
      tags: [videos]
      summary: Bookmark a video
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkVideoBody"
      responses:
        "200":
          description: Bookmark toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookmarkResult"
  /videos/bookmarks:
    get:
      operationId: listBookmarkedVideos
      tags: [videos]
      summary: List bookmarked videos for a user
      parameters:
        - in: query
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bookmarked videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"

  # --- DASHBOARD ---
...
components:
  schemas:
...
    Twin:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        label:
          type: string
          description: "Twin A or Twin B"
        name:
          type: string
        gender:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        colorTheme:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - label
        - name
        - colorTheme
        - createdAt
        - updatedAt
    CreateTwinBody:
      type: object
      properties:
        userId:
          type: string
        label:
          type: string
        name:
          type: string
        gender:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        colorTheme:
          type: string
      required:
        - userId
        - label
        - name
        - colorTheme
    UpdateTwinBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        birthdate:
          type: ["string", "null"]
        profilePicture:
          type: ["string", "null"]
        colorTheme:
          type: ["string", "null"]
    SleepEntry:
      type: object
      properties:
        id:
          type: integer
        twinId:
          type: integer
        type:
          type: string
          description: "nap or night"
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - twinId
        - type
        - startTime
        - createdAt
    CreateSleepEntryBody:
      type: object
      properties:
        twinId:
          type: integer
        type:
          type: string
          description: "nap or night"
        startTime:
          type: string
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
      required:
        - twinId
        - type
        - startTime
    UpdateSleepEntryBody:
      type: object
      properties:
        endTime:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
    SleepSummary:
      type: object
      properties:
        twinId:
          type: integer
        date:
          type: string
        dailyTotalMinutes:
          type: integer
        weeklyTotalMinutes:
          type: integer
        napCount:
          type: integer
        nightSleepMinutes:
          type: integer
      required:
        - twinId
        - date
        - dailyTotalMinutes
        - weeklyTotalMinutes
        - napCount
        - nightSleepMinutes
...
    Video:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
          description: "sleep, feeding, pregnancy, toddler, day-in-the-life, tips"
        sourceType:
          type: string
          description: "youtube, upload"
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - category
        - sourceType
        - url
        - createdAt
    CreateVideoBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        category:
          type: string
        sourceType:
          type: string
        url:
          type: string
        thumbnailUrl:
          type: ["string", "null"]
        durationSeconds:
          type: ["integer", "null"]
      required:
        - title
        - category
        - sourceType
        - url
    BookmarkVideoBody:
      type: object
      properties:
        userId:
          type: string
        bookmarked:
          type: boolean
      required:
        - userId
        - bookmarked
    BookmarkResult:
      type: object
      properties:
        videoId:
          type: integer
        userId:
          type: string
        bookmarked:
          type: boolean
      required:
        - videoId
        - userId
        - bookmarked
...